refactor and document wrangler dev
