+++ Cargo.toml
@@ -28,7 +28,6 @@ exitfailure = "0.5.1"
 failure = "0.1.5"
 flate2 = "1.0.7"
 fs2 = "0.4.3"
-futures = "0.3"
 futures-util = "0.3"
 http = "0.2.0"
 hyper = "0.13.1"
@@ -53,7 +52,7 @@ sha2 = "0.8.0"
 tempfile = "3.1.0"
 term_size = "0.3"
 text_io = "0.1.7"
-tokio = { version = "0.2", default-features = false, features = ["io-std", "time"] }
+tokio = { version = "0.2", default-features = false, features = ["io-std", "time", "macros"] }
 tokio-tungstenite = { version = "0.10.1", features = ["tls"] }
 toml = "0.5.5"
 url = "2.1.0"
