+++ src/commands/dev/gcs/headers.rs
@@ -1,5 +1,15 @@
+/// The preview service runs on cloudflareworkers.com and then sends parts
+/// of the incoming request to the `request` in the Workers Runtime
+///
+/// The way it does this is by prefixing the _real_ headers sent to
+/// and returned by the worker with this header prefix
 const HEADER_PREFIX: &str = "cf-ew-raw-";
-
+/// Any request headers sent to `wrangler dev` must be prefixed
+/// before sending it to the preview service
+/// and any response headers sent from the preview service
+/// that don't have the prefix must be removed
+/// and response headers that do have the prefix
+/// must have the prefix stripped
 use std::str::FromStr;
 
 use hyper::header::{HeaderMap, HeaderName};
@@ -7,15 +17,21 @@ use hyper::http::request::Parts as RequestParts;
 use hyper::http::response::Parts as ResponseParts;
 use hyper::http::status::StatusCode;
 
+/// modify an incoming request before sending it to the preview service
 pub fn structure_request(parts: &mut RequestParts) {
     prepend_request_headers_prefix(parts)
 }
 
+/// modify a response from the preview service before returning it to the user
 pub fn destructure_response(parts: &mut ResponseParts) -> Result<(), failure::Error> {
     set_response_status(parts)?;
     strip_response_headers_prefix(parts)
 }
 
+/// every header sent to `wrangler dev` must be prefixed
+/// before sending it along to the preview service
+/// so it is sent along to the Workers runtime and not
+/// consumed directly by the preview service
 fn prepend_request_headers_prefix(parts: &mut RequestParts) {
     let mut headers: HeaderMap = HeaderMap::new();
 
@@ -29,6 +45,12 @@ fn prepend_request_headers_prefix(parts: &mut RequestParts) {
     parts.headers = headers;
 }
 
+/// every header returned by the Worker is prefixed
+/// and headers without a prefix are specific to the preview service
+/// here we parse those headers and construct a new header map
+///
+/// discard headers without that prefix
+/// strip the prefix from real Workers headers
 fn strip_response_headers_prefix(parts: &mut ResponseParts) -> Result<(), failure::Error> {
     let mut headers = HeaderMap::new();
 
@@ -45,13 +67,16 @@ fn strip_response_headers_prefix(parts: &mut ResponseParts) -> Result<(), failur
     Ok(())
 }
 
+/// parse the response status from headers sent by the preview service
+/// and apply the parsed result to mutable ResponseParts
 fn set_response_status(parts: &mut ResponseParts) -> Result<(), failure::Error> {
     let status = parts
         .headers
         .get("cf-ew-status")
         .expect("Could not determine status code of response");
-    // status will be "404 not found" or "200 ok"
-    // we need to split that string to create hyper's status code
+
+    // `status` above will be "404 not found" or "200 ok"
+    // we need to parse that string to create hyper's status code
     let status_vec: Vec<&str> = status.to_str()?.split(' ').collect();
     parts.status = StatusCode::from_str(status_vec[0])?;
     Ok(())
